Select the authentication-provider base on a property
The authentication provider can be selected using a maven profile or a
environment property.
Property is authentication.mode:
 * `-Dauthentication.mode=ba` -> Local in-memory authentication (default
       if no other is passed)
 * `-Dauthentication.mode=ldap` -> Remote LDAP server
 * `-Dauthentication.mode` -> Local LDAP server.

You can use a profile instead a property:
 * By default local in-memory authentication is used.
 * `-P ldap` -> Remote LDAP server
 * `-P ldap-dev` -> Local LDAP server.

For the local in-memory authentication the users are defined in the file
users.properties with this format:
`username=password,grantedAuthority[,grantedAuthority][,enabled|disabled]`

The first entry of the file is filtered to use the user defined in
pom.xml.